Fix for circular dependecy in imports. (#1933)
* Move popOverOverlayPosition into new file
