+++ tests/utils/utils.test.jsx
@@ -496,16 +496,4 @@ describe('Utils.isKeyPressed', function() {
         const key = ['k', 2221];
         expect(Utils.isKeyPressed(event, key)).toEqual(true);
     });
-
-    test('Should return placement position for overlay based on bounds, space required and innerHeight', function() {
-        const targetBounds = {
-            top: 400,
-            bottom: 500,
-        };
-
-        expect(Utils.popOverOverlayPosition(targetBounds, 1000, {above: 300})).toEqual('top');
-        expect(Utils.popOverOverlayPosition(targetBounds, 1000, {above: 500, below: 300})).toEqual('bottom');
-        expect(Utils.popOverOverlayPosition(targetBounds, 1000, {above: 450})).toEqual('bottom');
-        expect(Utils.popOverOverlayPosition(targetBounds, 1000, {above: 600})).toEqual('left');
-    });
 });
